chore: Split solver
